+++ web/src/views/SignUpView.vue
@@ -1,4 +1,4 @@
- <template>
+<template>
   <router-link to="/">
     <img class="button-back" src="../assets/angulo-esquerdo.png" />
   </router-link>
@@ -8,13 +8,8 @@
     progressing</p>
 
   <div class="login">
-    <form>
-      <input type="email" class="form-input" placeholder="Email" id="inpuEmail" v-model="userData.email">
-      <input type="password" class="form-input" placeholder="Password" id="inputPassword" @keyup.enter="userLogin()"
-        v-model="userData.password">
-    </form>
+    <SignUpInputComponent/>
 
-    <button class="login-button" @click.prevent="userLogin()">Login</button>
     <small>Or Login In Using</small>
     <button class="login-with-google">
       <img src="https://cdn.jsdelivr.net/gh/devicons/devicon/icons/google/google-original.svg" />
@@ -22,7 +17,7 @@
     </button>
 
     <router-link to="/sign-in" class="link">
-      <p>Not a member? <a class="link">Create a user</a> </p>
+      <p>Not a member? <a class="link">Create a user</a></p>
     </router-link>
 
     <LoaderComponent />
@@ -30,55 +25,14 @@
 </template>
 
 <script lang="ts">
-import axios from 'axios'
-import { toast } from 'vue3-toastify';
-import 'vue3-toastify/dist/index.css';
-
 import LoaderComponent from '../components/LoaderComponent.vue';
+import SignUpInputComponent from '../components/SignUpInputComponent.vue';
 
 export default {
   name: "Login",
   components: {
-    LoaderComponent
-  },
-  data() {
-    return { userData: { email: "", password: "" } }
-  },
-  methods: {
-    async userLogin() {
-      try {
-        const loader = document.getElementById("loader_race_by") as HTMLElement
-        loader.style.display = "flex"
-
-        await axios.post("http://localhost:8080/api/user/sign-up", this.userData).then(res => {
-          const getFirstName: string[] = res.data.name.split(" ")[0].toLowerCase()
-
-          this.$router.push(`/task-s/${getFirstName}/${res.data.id}`)
-          this.getUserData(res.data.id)
-
-        })
-      } catch (e) {
-        toast.error('ERROR - Invalid password or email', {
-          autoClose: 3000,
-          theme: 'dark'
-        })
-
-        const inputEmail = document.getElementById("inpuEmail") as HTMLElement
-        const inpuPassword = document.getElementById("inputPassword") as HTMLElement
-
-        const color: string = "#D34242"
-        let inputs: Array<HTMLElement> = [inputEmail, inpuPassword]
-
-        for (let input of inputs) {
-          input.style.borderColor = color
-        }
-      }
-    },
-    async getUserData(userId: number) {
-      axios.get(`http://localhost:8080/api/all/users/${userId}`).then(res => {
-        console.log(res.data)
-      })
-    }
+    LoaderComponent,
+    SignUpInputComponent
   }
 }
 </script>
@@ -91,27 +45,6 @@ export default {
   border-radius: .5rem;
 }
 
-.form-input {
-  font-size: 1rem;
-
-  font-weight: 400;
-  width: 100%;
-
-  margin-top: 1.5rem;
-
-  line-height: inherit;
-
-  padding: 0.75rem 1.25rem;
-
-  border-radius: .5rem;
-  border: 1px solid #ebebeb6c;
-  background: transparent;
-
-  color: white;
-
-  transition: all 0.2s ease-in-out;
-}
-
 .passwords input {
   width: 80%;
 }
@@ -120,9 +53,13 @@ export default {
   margin-left: .5rem;
 }
 
-.passwords {
+.passwords,
+.login-with-google {
   display: flex;
   flex-direction: row;
+}
+
+.passwords {
   justify-content: space-between;
 }
 
@@ -136,8 +73,6 @@ export default {
 }
 
 .login-with-google {
-  display: flex;
-  flex-direction: row;
   justify-content: center;
 
   gap: 1rem;
@@ -150,18 +85,6 @@ export default {
   border: none;
 
   cursor: pointer;
-}
-
-.login-button {
-  outline: none;
-  border: none;
-  cursor: pointer;
-
-  background: #8257E5;
-  color: white;
-
-  border-radius: .5rem;
-  transition: .3s;
 }
 
 .login-with-google img {
